installd: insert one backdoor to support boot via nfs
when boot via nfs, we need the install usr on nfs server to have the
chown permission as it does on android. but that's no easy to do with
setting of nfs. so we will make the install to run as root via the
property hack.installd.for.nfs set to 1 when want to boot via nfs

Change-Id: I101849f9b7f55b9e8dd657ae321cf9341ddc7d02
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>